refactor(core): create stateManager service
Before: error creation by ESLint because the modelManager has to many statements.
Solution: divide modelManager into two entities: modelManager for schema and model and stateManager for state.

Closes #145
