+++ .eslintrc.yml
@@ -164,7 +164,7 @@ rules:
   max-len: 0
   max-nested-callbacks: 0
   max-params: 0
-  max-statements: [2, 50]
+  max-statements: [2, 30]
   new-cap: 0
   new-parens: 0
   newline-after-var: 0
