+++ src/app/ui/forms/test/test.directive.js
@@ -27,7 +27,7 @@ function avTest(schemaForm) {
     return directive;
 }
 
-function Controller($scope, $translate, events, modelManager) {
+function Controller($scope, $translate, events, modelManager, stateManager) {
     'ngInject';
     const self = this;
     self.modelName = 'test';
@@ -59,7 +59,7 @@ function Controller($scope, $translate, events, modelManager) {
         // First we broadcast an event so all fields validate themselves then we validate the model to update
         // summary panel
         $scope.$broadcast('schemaFormValidate');
-        modelManager.validateModel(self.modelName, $scope.activeForm, $scope.form[0].tabs);
+        stateManager.validateModel(self.modelName, $scope.activeForm, $scope.form[0].tabs);
     }
